+++ src/main/java/org/jperera/refactoringdifferentlevels/PictureRepository.java
@@ -2,5 +2,5 @@ package org.jperera.refactoringdifferentlevels;
 
 import org.springframework.data.repository.CrudRepository;
 
-public interface PictureRepository extends CrudRepository<Picture, Long> {
+public interface PictureRepository extends CrudRepository<Picture, Integer> {
 }
